[DSPlayer] merged from trunk revs r34353:34770
git-svn-id: https://xbmc.svn.sourceforge.net/svnroot/xbmc/branches/dsplayer@34775 568bbfeb-2a22-0410-94d2-cc84cf5bfa90

+++ tools/EventClients/Clients/WiiRemote/WiiUse_WiiRemote.cpp
@@ -182,9 +182,9 @@ float smoothDeg(float oldVal, float newVal)
   return(newVal);
 }
 
-long getTicks(void)
+int32_t getTicks(void)
 {
-  long ticks;
+  int32_t ticks;
   struct timeval now;
   gettimeofday(&now, NULL);
   ticks = now.tv_sec * 1000l;
@@ -194,7 +194,7 @@ long getTicks(void)
 
 int main(int argc, char** argv) 
 {
-  long timeout = 0;
+  int32_t timeout = 0;
   bool connected = 0;
   wiimote** wiimotes;
   wiimotes =  wiiuse_init(MAX_WIIMOTES);
